Move `loadMetadataForCard` to its own file
